+++ src/main/java/com/hod/project/hive/test/service/mapper/TestMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.hod.project.hive.mapper.TestMapper">
+<mapper namespace="com.hod.project.hive.test.service.TestMapper">
     <select id="findSomething" resultType="string">
         SELECT uuid() AS id
     </select>
@@ -15,7 +15,7 @@
         SELECT 1234 AS id
     </select>
 
-    <select id="findObject" resultType="com.hod.project.hive.entity.TestObject">
+    <select id="findObject" resultType="com.hod.project.hive.test.entity.TestObject">
         SELECT 1234 AS id
     </select>
