+++ src/main/java/com/hod/project/hive/test/vo/RequestParams.java
@@ -1,4 +1,4 @@
-package com.hod.project.hive.common.vo;
+package com.hod.project.hive.test.vo;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
