+++ src/main/resources/application.properties
@@ -9,4 +9,5 @@ spring.datasource.username=sa
 spring.datasource.password=
 
 mybatis.configuration.map-underscore-to-camel-case=true
-mybatis.mapper-locations=classpath:mapper/*.xml:wq:
+mybatis.type-aliases-package=true
+mybatis.mapper-locations=**/mapper/*.xml
